# src/swagger.yaml
openapi: 3.0.3
info:
  title: ORGiD migration backend server
  description: >-
    This is the interface documentation for the RESTful API interface for
    backend services of ORGiD migration backend server
  version: 0.0.1
servers:
  - url: http://localhost:3333/api
    description: Local Testing server
  - url: https://api.test-orgid-migration.windingtree.com/api
    description: Staging server
  - url: https://api.orgid-migration.windingtree.com/api
    description: Production server
paths:
  /health:
    get:
      description: Service health report
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clean:
    post:
      description: System reset (on test only)
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /file:
    post:
      description: Files upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedFile'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /fileUri:
    post:
      description: Files upload by its URI
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ApiFileUriParams'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedFile'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dids/{owner}:
    get:
      description: Returns list of owned ORGiDs DIDs
      parameters:
        - in: path
          name: owner
          required: true
          schema:
            $ref: '#/components/schemas/EthAddress'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dids'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /did:
    get:
      description: Returns ORG.JSON of the DID
      parameters:
        - in: query
          name: did
          required: true
          schema:
            $ref: '#/components/schemas/Did'
      responses:
        '200':
          description: OK
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/OrgJsonString'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /request:
    post:
      description: Creation of migration requests job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatus'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /request/id/{id}:
    get:
      description: Returns request status by Id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/RequestId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatus'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /request/did:
    get:
      description: Returns request status of the did
      parameters:
        - in: query
          name: did
          required: true
          schema:
            $ref: '#/components/schemas/Did'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatus'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      description: Objet describing an error
      type: object
      properties:
        message:
          description: Error message
          type: string
          example: Something went wrong
      required:
        - message
    Health:
      description: Health response
      type: object
      properties:
        time:
          description: Current server time
          type: string
          format: date-time
          example: '2022-11-16T22:33:49.333+01:00'
        commit:
          description: Actual git commit hash
          type: string
          example: '7edb2b9e2204f908d1b03f2c2872b39f29b31bc5'
        redis:
          description: Redis DB connection status
          type: string
          example: 'ready'
      required:
        - time
        - commit
        - redis
    EthAddress:
      description: EOA
      type: string
      pattern: ^0x[a-fA-F0-9]{40}$
      example: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266'
    Did:
      description: ORGiD DID
      type: string
      pattern: ^did:orgid:([0-9])*(?::)?(0x[a-fA-F0-9]{64})+$
      example: did:orgid:77:0xeae18b4ccf6dfd743dc1738c6547bc829c3384bcc2b48f98f0c9a49c5c67b2be
    DidWithState:
      description: List of ORGiDs DIDs
      type: object
      properties:
        did:
          $ref: '#/components/schemas/Did'
        newDid:
          $ref: '#/components/schemas/Did'
        state:
          $ref: '#/components/schemas/RequestState'
      required:
        - did
        - state
    Dids:
      description: List of ORGiDs DIDs
      type: array
      items:
        $ref: '#/components/schemas/DidWithState'
    ApiOwnerParams:
      description: In-path API /owner/{address} parameters
      type: object
      properties:
        owner:
          $ref: '#/components/schemas/EthAddress'
      required:
        - owner
    ApiDidParams:
      description: In-path API /did/{did} parameters
      type: object
      properties:
        did:
          $ref: '#/components/schemas/Did'
      required:
        - did
    ApiRequestParams:
      description: In-path API /request/{id} parameters
      type: object
      properties:
        id:
          $ref: '#/components/schemas/RequestId'
      required:
        - id
    ApiFileUriParams:
      description: In-body API /fileUri parameters
      type: object
      properties:
        file:
          description: File URI
          type: string
      required:
        - file
    RequestId:
      description: Migration request Id
      type: string
    MigrationRequest:
      description: An ORGiD migration request
      type: object
      properties:
        did:
          $ref: '#/components/schemas/Did'
        chain:
          description: Chain Id
          type: number
          enum: [5, 77, 100, 137]
        orgIdVc:
          description: Serialized ORGiD VC
          type: string
      required:
        - did
        - chain
        - orgIdVc
    RequestState:
      description: An ORGiD migration request progress state
      type: string
      enum:
        - ready
        - requested
        - progress
        - failed
        - completed
    RequestStatus:
      description: An ORGiD migration request status
      type: object
      properties:
        id:
          $ref: '#/components/schemas/RequestId'
        timestamp:
          description: Timestamp when the request was created
          type: number
        did:
          $ref: '#/components/schemas/Did'
        newDid:
          $ref: '#/components/schemas/Did'
        state:
          $ref: '#/components/schemas/RequestState'
      required:
        - id
        - timestamp
        - did
        - newDid
        - state
    OrgJsonString:
      description: Serialized ORG.JSON doc
      type: string
    UploadedFile:
      description: File upload API response
      type: object
      properties:
        url:
          description: Url of the uploaded file
          type: string
      required:
        - url